requirements.txt
python-telegram-bot==20.3
python-dotenv==1.0.0
aiohttp==3.8.4